feat(zdo): handle APS ACKs fragments and MatchDesc
Dispatch staged APS ACK and fragment events from poll_aps and add the ZDP MatchDescReq client helper.

+++ zigbee/src/apl/descriptors/simple_descriptor.rs
@@ -67,10 +67,10 @@ mod tests {
             APPLICATION_INPUT_CLUSTER_COUNT
         );
         assert!(!simple_descriptor.application_input_cluster_list.is_empty());
-        for i in 0..APPLICATION_INPUT_CLUSTER_COUNT as usize {
+        for i in 0..APPLICATION_INPUT_CLUSTER_COUNT {
             assert_eq!(
-                simple_descriptor.application_input_cluster_list[i],
-                (i + 1) as u8
+                simple_descriptor.application_input_cluster_list[usize::from(i)],
+                i + 1
             );
         }
         assert_eq!(
@@ -78,10 +78,10 @@ mod tests {
             APPLICATION_OUTPUT_CLUSTER_COUNT
         );
         assert!(!simple_descriptor.application_output_cluster_list.is_empty());
-        for i in 0..APPLICATION_OUTPUT_CLUSTER_COUNT as usize {
+        for i in 0..APPLICATION_OUTPUT_CLUSTER_COUNT {
             assert_eq!(
-                simple_descriptor.application_output_cluster_list[i],
-                (i + 2) as u8
+                simple_descriptor.application_output_cluster_list[usize::from(i)],
+                i + 2
             );
         }
     }
@@ -115,10 +115,10 @@ mod tests {
             APPLICATION_INPUT_CLUSTER_COUNT
         );
         assert!(!simple_descriptor.application_input_cluster_list.is_empty());
-        for i in 0..APPLICATION_INPUT_CLUSTER_COUNT as usize {
+        for i in 0..APPLICATION_INPUT_CLUSTER_COUNT {
             assert_eq!(
-                simple_descriptor.application_input_cluster_list[i],
-                (i + 1) as u8
+                simple_descriptor.application_input_cluster_list[usize::from(i)],
+                i + 1
             );
         }
         assert_eq!(simple_descriptor.application_output_cluster_count, 0);
@@ -156,10 +156,10 @@ mod tests {
             APPLICATION_OUTPUT_CLUSTER_COUNT
         );
         assert!(!simple_descriptor.application_output_cluster_list.is_empty());
-        for i in 0..APPLICATION_OUTPUT_CLUSTER_COUNT as usize {
+        for i in 0..APPLICATION_OUTPUT_CLUSTER_COUNT {
             assert_eq!(
-                simple_descriptor.application_output_cluster_list[i],
-                (i + 2) as u8
+                simple_descriptor.application_output_cluster_list[usize::from(i)],
+                i + 2
             );
         }
     }
